Исправлен метод связывания узлов списка. Ссылка на хвост не обновлялась при добавлении второго элемента.
Исправлено местоположение класса Node. Изначально запутался в ТЗ, подумал, что нужно его создать в том же документе, что и менеджер истории.
